fixed bugs associated with drag/drop interface.
First bug: drag/drop was not working. This was a bug in
onPlaylistSongMoved.

Second bug: dragging a song to the end of the playlist duplicated it
visually, but not on the server. Hacky fix is to delete all
non-draggable songs after the playlist is refreshed.

Third bug: clicking on the drag bar but not dragging disabled dragging
for the song and prevented further dragging. Ended up clearing the
dragged element after mouseup in the Sortable.js code.